make location tab active after import. refactoring and cleanup `[refactoring]` `[bug]`
